Use path in TMPDIR for schroot's overlay
By default, overlay path for schroot is /var/lib/schroot/union/overlay
that leads to the conflict with docker build.

Docker itself also uses overlayfs for the layers and we come to
situation with 'overlayfs over overlayfs' that is not supported
by the driver.

This solution places overlay path under the build/tmp so gitlab-runner
configuration changes or using external volumes in docker are not
required anymore.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>