vm-img: fix clock sync issues in vmware template
The time inside a default Isar VM is UTC but the RTC provided by vmware
could be non UTC so one would see a multi-hour clock offset between
guest and host. The offset would depend on the timezone used on the
host, and also on where that host keeps its timezone.

This patch fixes the issue, both for Windows hosts (timezone in RTC) and
for Linux (no timezone in RTC) hosts.

Signed-off-by: Henning Schild <henning.schild@siemens.com>