buildchroot: delay expansion of variable BUILDCHROOT_DIR
The current implementation expands ${BUILDCHROOT_DIR_HOST|TARGET} during
assignment to BUILDCHROOT_DIR, which leads to absolute paths in signatures.
This can be observed in IMAGE_CMD_* and CONVERSION_CMD_*, and is bad for caching.

Thus, we delay expansion, and due to the way the do_image_* tasks are created,
the paths in our signatures will be relative to ${TMPDIR} (no expansion).

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>