make debug symbols of kernel build relocatable
This patch adds support to inject custom KCFLAGS into
the linux-custom kernel build.
Using this infrastructure, we add the debug-prefix-map
flag to strip off the absolute prefix of the sources location.
This makes the debug information relocatable.

By that, we also fix low ccache hit rates in combination with sbuild
due to the random builddir that sbuild uses.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Anton Mikanovich <amikan@ilbers.de>