ccache: make sure ccache is really not used even when installed
As soon as ccache is installed the build system meson will jump on it
and tries to use it. But it is not really usable when USE_CCACHE is not
enabled, the build will run into permission problems creating cache
files which anyhow will be ditched with the sbuild chroot.

Could be considered a bug in meson that it will use ccache if it is
installed, or it could be a bug in isar where ccache can not be used
inside the sbuild chroot ... while it is installed there.

Simply disabling it inside the chroot when disabled for that package
works well.

Signed-off-by: Henning Schild <henning.schild@siemens.com>