fix on-first-boot for systemd >= 247
This patch fixes the bug that /etc/machine-id is never
generated when running from within an initrd with an ro rootfs.

In this case, the magic value "uninitialized" has to be used
instead of deleting the file.

For older systemd versions, we set it to the empty string
which is non-documented, but works as identified
in various tests on debian buster (systemd version 241).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>