wic: bootimg-efi-isar: Sync with latest upstream changes
Specifically brings in DTB support for grub and systemd-boot.

It also changes the generation of unified kernel images in a way that
is definitely not compatible with Isar. Not sure if this feature was
used before (isar-cip-core uses EFI Boot Guard for that purpose), but
it's still patched here to not call non-existing objcopy binaries, just
in case.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>