ignore wic tasks in isar-sstate lint
This patch improves the isar-sstate linter.
It adds support to ignore well-known tasks that
are not cacheable anyway (e.g. rootfs_wicenv and image_wic).

This sensible default can be overwritten by using --pedantic
to also check for cachability issues in these tasks.
This helps to spot transitive caching issues happening when
performing image-in-image builds, but requires special preparation
of the calling environment (mainly no absolute paths in BBLAYERS).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>