unify ubuntu-focal.conf with debian config files
This patch reworks the ubuntu-focal.conf distro config file.
It unifies the pattern used for DISTRO_APT_SOURCES with the
pattern used in debian-common.conf.

By that, the BASE_DISTRO variable is used to select the corresponding
apt sources.
This makes it possible to derive custom distros based on ubuntu as well.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>