image: remove / entry from fstab template
An fstab entry for / should usually not be required. It just makes a
rootfs less portable (USB vs. NFS, vs. SSD). The kernel is the one
that should know which fs to mount and how. That is done via a bootloader.

Should additional mount options be needed for a concrete image, an
imager like wic might add an entry. i.e. for adding fspassno or
x-systemd.growfs

So to make the rootfs more generic and to allow imagers to add their
individual / lines, instead of having to modify an existing one, remove
the line from the template.

Signed-off-by: Henning Schild <henning.schild@siemens.com>