Exit bootstrap with error if apt-get update fails
If `apt-get update` can't reach one ore more mirrors listed in
sources.list, it consider this as 'transient' error and doesn't
produce non-zero error code.

So, `do_bootstrap` finishes with no errors in this case and we don't
have proper apt database here and face with errors in the later tasks,
e.g. `rootfs_install_pkg_download` in buildchroot.

The patch adds an option to `apt-get update` that makes it exit with
non-zero code for trancient failures.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>