fix systemd-boot for non x64 targets
With the recently introduced cross-imaging, all imager dependencies are
installed for the build architecture but not the target / distro
architecture. This is correct for tools, but not for architecture
specific files that are copied into the target rootfs.
By that, arm64 targets got bootx64.efi loaders installed, making it
impossible to boot the image.

For grub, we already have architecture specific support based on package
names. This patch introduces the same for systemd-boot. Here, the
architecture is not encoded in the package name, but in the architecture
field. To avoid conflicts with other packages in the buildchroot-host,
from bookworm on, we only depend on systemd-boot-efi,
as this package provides the efi stubs and is co-installable for
multiple architectures. For older releases, this is not possible but
depending on systemd works as well.

In theory this could conflict with a build-arch systemd installed into
the buildchroot-host, but as almost all dependencies are now installed
using sbuilder and schroot, this case is very unlikely.
In this case, a user still can disable cross for the imaging.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>