set rdeptask for do_deploy_deb
Isar currently supports setting build-dependencies between recipes with
DEPENDS, and runtime/installation dependencies with DEBIAN_DEPENDS.

However, the bitbake mechanism of RDEPENDS, for runtime dependencies
between recipes, is not supported. So currently, if we model a runtime
dependecy between recipes, we need to use DEPENDS, which at the same
time enforces a build dependency, i.e., restricts parallel execution
of builds.

To enable the use of RDEPENDS in Isar, we set the task (deploy_deb)
that should be linked by those dependencies.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>