Separate fast and full groups of run tests.
Before, "full" CI build executed both "full" and "fast" subsets
of run tests. So, it was impossible to add run test for a target
that existed only in "fast" group.

Now, run tests can be individually added for both "fast" and "full"
subsets and they won't depend on each other.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>