sbuild: Allow downgrades when installing dependencies
If the package we are building depends on previously built custom
package, but sbuild chroot happens to have it already installed,
we need to allow downgrades explicitly.

Related commits: 5367ba74d9f2 and 77e62c1e9bbd (for buildchroot).

Actually, the issue is not so typical for sbuild (in comparisong to
previously used buildchroot) while sbuild base chroot keeps the bare
minimum of preinstalled packages.

Anyway, the fix might be needed in case we want to recompile some
essential package and use it as a dependency for another one.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>