linux install.tmpl: added CONFIG_HAVE_OBJTOOL in OR for objtool
linux install.tmpl: added OR with CONFIG_HAVE_OBJTOOL for objtools
a fix for newer kernels (since 5.19 for sure) to include objtool

The newer kernels (e.g. v5.19) when compiled and packaged by an ISAR
do not include the objectool anymore in linux headears package. This
obviously breaks any activity which expects to use that binary
executable.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>