example-raw: fix postinst script dependency on adduser
Starting from Nov 2022, package `adduser` is not preinstalled
in bullseye when the distro debootstrapped.

This caused example-raw `postinst: 6: addgroup: not found` error
when installing `container-amd64-bookworm` image.

The problem was not reproduced on `qemuamd64-bookworm` while `adduser`
exist in the image as a dependency of some other package.

Fix the problem by adding explicit dependency to `adduser` package.
Also, we don't need dependency on `passwd` anymore while `adduser`
already depends on it in all distros.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>