fix apt-retry spec used by apt running in rootfs
This patch fixes a bug introduced in d349b7 when moving
the apt options from the apt cli to a config file:
The configuration to let apt retry failed downloads belongs to the
'Acquire' group but not to the 'APT' group of apt.conf.
Previously this had been placed in the invalid group and hence did
not have any effect.

According to the manpage of apt.conf, the new location is at least
valid from debian stretch on and from ubuntu bionic on. By that,
all distros supported by ISAR are covered. The same configuration
entry can also be seen in the debrebuild script in /usr/bin.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>