add interface to passthrough env vars to sbuild
This patch introduces the bb variable SBUILD_PASSTHROUGH_ADDITIONS.
All variable names that are added to it are forwarded into the sbuild
environment and also excluded from the bb signatures of the task.

By adding this interface, we provide a mechanism to inject external
mirrors for packages that need to download components at build time.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>