add support for DEB_BUILD_OPTIONS in linux-module
This patch reworks how concurrency is induced in the build of a custom
kernel module. Previously it was controlled via a hard-coded
DEB_BUILD_OPTIONS environment variable in the rules file. This made it
impossible to inject other build profiles from the calling environment.
In addition, the variable is only intended to be set from the outside,
not within a rules file.

Since debhelper 10, parallel builds are the default anyways, but when
overwriting the build target, manual passing of the argument is
required. Here, we now reference the standard PARALLEL_MAKE variable
which is nowadays used to control the concurrency within builds and is
also set by bitbake.

As the only relevant task for concurrency is the build of the modules,
we just apply it there.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>