fix: only append kargs and extra_kargs if set
This fixes an issue when no KARGS and no EXTRA_KARGS where set, leading
to an empty string after -append. This happens for instance when booting
a WIC image instead of performing a direct kernel boot.

Instead of always adding -append, we only add it if at least on of the
variables is set.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>