sbuild-chroot: Pre-install cross-compiler and libc6-dev
This saves build setup time by pre-installing not only build-essential
but - for cross and compat-arch cases - also the cross-compiler as well
as libc6-dev.

The latter is coming via build-essential in the native case, it's not
installed by crossbuild-essential but sbuild is pulling it. We can
install libc6-dev:${DISTRO_ARCH} unconditionally as this becomes a not
in the native case.

Things brings sbuild on eye-level with buildchroot regarding the initial
build environment, even goes beyond as buildchroot used to install only
libc6:<target-arch>.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>