sbuild-chroot: Pre-install also libcpp-*-dev and libstdc++-*-dev
Basically the same story as for libc6-dev: gcc and g++ pull those in for
the native case, but the cross-compilers will only pull the respective
dev-*-cross packages which do not help to fulfill build dependencies.

As those package names depend on the compiler version, introduce a new
distro configuration variable DISTRO_GCC. If it is set, it shall contain
the default gcc version number. If it is not set, the pre-installation
optimization is skipped by sbuild-chroot.

DISTRO_GCC is set for all in-tree distro versions except for sid as it
is too much of a moving target there.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>