image: make sure do_rootfs_finalize can run multiple times
There have been several places that would not work twice in a row. All
the test && not covering || have been moved to if. The rm at the end
tried to remove a file that was just moved. And that mv would not work a
second time because its source would be gone.

Signed-off-by: Henning Schild <henning.schild@siemens.com>