remove apt cache dirs from final rootfs
In 0c177a4b an improvement was added to not add the apt cache to the
sstate cache. This lead to reproducibility issues in case the rootfs is
provided by the sstate cache, compared to a newly assembled one.

As these folders are created on-the-fly when running apt, we simply
remove them from the final image.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>