sources.list: add non-free-firmware repository
The Debian project has taken the decision in 2022-10 to
create a new repository component non-free-firmware, and
include its content on installation media for the upcoming
Debian release bookworm. This has to be reflected in our
sources.list from bookworm on.

While the non-free-firmware repository contains only the
firmware parts, we could remove non-free as the intention
of adding non-free was just to get the non-free firmware
packages. However, still many firmware packages are not in
the correct repository, so better keep non-free for now.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>