fix race-cond between default and custom initrd
This patch fixes a data race happening when building a custom initrd.
Previously, both custom and default initrds were deployed to the image
deploy dir. The race is fixed by conditionally deploying either the
custom or the default one. For that, we introduce a new variable
INITRD_DEPLOY_FILE which provides the name of the initrd in the deploy
directory. The existing INITRD_IMAGE variable is defaulted to the empty
string and used to control if a custom initrd is requrested. Only if
this variable is empty, the default one is deployed.

Uladzimir Bely: Drop redundant expand=True parameter from getVar

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Uladzimir Bely <ubely@ilbers.de>