vm-img: virtualbox: update ovf template
These templates are by enlarge taken from exported VMs, so it is really
hard to change only one thing at a time and ensuring things will still
work as expected. So this commit does multiple things at a time.

It repairs the ofv by taking the StorageControllers into the Hardware
block.
It switches the default disk from IDE to virtio.
It adds a USB host controller and pointer device.

Signed-off-by: Henning Schild <henning.schild@siemens.com>