fix xattrs of rootfs when using sstate cache
When creating the sstate cache artifact, the extended file attributes
need to be stored in the tar archive as well. Otherwise, capabilities on
files are lost when the artifact from the cache is used.

Note, that for tar versions shipped with debian bullseye and bookworm,
despite of the documentation, both --xattrs and --xattrs-include='*' is
required to store all extended attributes.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>