make compressed images better syncable
For compression formats that support these options (currently only zstd),
we enable to include the sparseness information into the archive. By that,
disk IO on decompression of large zero blocks is speedup and storage usage
is decreased. Usually, this is handled automatically by the compression
tools, but not when sending output to stdout.

In addition, we add the rsyncable flag to reduce the number
of changed blocks in the output when only changing some input blocks.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>