add zchunk compression support
This patch adds support to compress images with zchunk. The compression
format is optimized for delta transfers, by ensuring that small changes
on the input also only invalidate some blocks of the output.

There are not yet many configuration options available, but once
available, they can be added to the ZCK_DEFAULTS variable.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>