meta-isar: add support for ubuntu jammy
This patch adds the ubuntu jammy distribution and extends the qemuamd64
and qemuarm64 machine configs to support that version. Multiconfig
entries are added as well. As ubuntu jammy uses zstd compression for
the .deb files, the bootstrapping requires host dpkg >= 1.21, which is
shipped from debian bookworm on. This note is added to the user docs.
No test is added because of the special host tooling requirements.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>