meta-isar: Cleanup machine configs
Some stuff from IMAGE_INSTALL / IMAGE_PREINSTALL like `sshd-regen-keys`
or `expand-on-first-boot` should not be part of machine configs.

Move it to the isar-image-ci configuration for the targets that
we tested. If a downstream relays on these packages for some hardware
targets they should adjust their `local.conf` or image recipe.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>