inc files: append local files path to FILESPATH
FILESEXTRAPATHS is meant to allow bbappend files to overwrite files from
their recipe. For this reason the 'normal' pattern in bbappend files is
to prepend the local file path to FILESEXTRAPATHS, which results in
files from the local file path to be preferred over files from the
original recipe, or other, lower prioritized bbappend files.

The `.inc` files are meant to be included by other recipes (`.bb`)
files. So it should still allow the recipe (`.bb` file) to overwrite
files its files that where added via the default recipes `FILESPATH`
search path.

If a `.inc` file use `FILESEXTRAPATHS:prepend` to add its local file
path to `FILESPATH`, then the recipe, that includes the `.inc` file
would have to use `FILESEXTRAPATHS:prepend` as well, after including the
`.inc` file in order to overwrite the local files from the `.inc` file.

However this is only a workaround of the issue that originates from the
`.inc` file using the `FILESEXTRAPATHS` variable.

`.inc` should append its local file search path to `FILESPATH`, so that
it allows other recipes to overwrite any file from it via its default
`FILESPATH`, without requiring an additional workaround.

This patch replaces all occurences of
	FILESEXTRAPATHS:prepend := "${FILE_DIRNAME}/files:"
with
	FILESPATH:append := ":${FILE_DIRNAME}/files"
in the `.inc` files.

Signed-off-by: Claudius Heine <ch@denx.de>
Signed-off-by: Uladzimir Bely <ubely@ilbers.de>