cibuilder.py: Support custom arguments passing to CI scripts
With this change, it becomes possible to add custom arguments
to the scripts we want to run remotely on the target.

This means that remote script execution process is now consist
of two parts: delivering the script on board and running it
with custom arguments.

An interface of the call doesn't change just allowing to add
space-separated arguments:

```
self.vm_start(script='test_script arg1 arg2 .. argN')
```

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>