fix(u-boot-script): use correct ramdisk size
When booting the kernel via a boot* command, the ramdisk is passed as
address : size. However, the size pointed to the temporary variable
filesize, which denotes the filesize of the last load. When combining
with DT overlays, the overlays are loaded later and by that overwrite
the filesize variable, resulting in a partially passed initrd.

To fix this, we assign the size immediately after loading it to an env
variable and use that when booting.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>