use builtin DT for nanopi-neo target
The upstream kernel still misses the __symbol__ information in the
device tree. By that, overlays cannot be applied (DT was compiled
without -@). However, all u-boot internal device trees are compiled with
symbol information. By that, just use the one from u-boot.
Note, that the mmclbk entry changed in this device tree (mmclbk2
instead of mmclbk0). This change is reflected in the wks file. An
alternative option would be to use --use-uuid.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>