meta/bitbake.conf: add MACHINEOVERRIDES and DISTROOVERRIDES
Replacing MACHINE and DISTRO in the OVERRIDES variable with
MACHINEOVERRIDES and DISTROOVERRIDES allows to configure machine or
distro compatibilities by prepending to those variables.

This is smilar to how OE handles this.

In Isar, not setting the DISTRO variable is not supported, therefore we
don't need a inline python function for it.

Signed-off-by: Claudius Heine <ch@denx.de>