testsuite: Rework cache populating of sstate-cache in SstateTest
In SstateTest class we run first build to populate sstate-cache and
the following builds to check if it's used or not.

Since both first and following builds are started from the same isar
location, this approach doesn't allow to find possible issues that are
sensitive to this path.

This patch moves sstate-cache populating to a separate test so that
it uses a different isar location, e.g.:

1st build: "./isar-sstate" => "./isar-sstate/../sstate-cache";
2nd build: "."             => "./sstate-cache".

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>