add EFI boot version of nanopi-neo target
This patch adds the nanopi-neo-efi machine that implements the EFI boot
chain on the nanopi-neo board. This is only supported from debian
bookworm on, as a more recent u-boot version is required (2021.01 which
is shipped in bullseye does not work).

The EFI boot pattern later can be used to support mechanisms like
swupdate.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>