meta: optee-client: Generalize and update shutdown rule
First, naming changed somewhere between kernel 5.10 and 6.1 so that
optee-ta-bc50d971... is now optee-ta-supp-bc50d971...

And then we should account for possibly more devices that need to be
unbound. Use POSIX-compliant shell code for this.

The generic loop allows to drop the explicit unmounting of efivarfs
because the corresponding driver will already get unbound.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>