crossvars: Adjust logic to account for -native package builds in non-cross setups
When selecting a -native version of a package while cross-building is
disabled for it, we will use the target build env instead. That may
work, just slowly, or even fail miserably.

Extend the build env selection logic to take PACKAGE_ARCH into account
as well. It defaults to DISTRO_ARCH for non-package recipe, thus should
works fine for those as well.

Reported-by: Cedric Hombourger <cedric.hombourger@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>