bitbake: backport fetch2/aws: forward env-vars used in gitlab-ci K8s
The patch "fetch2/aws: forward env-vars used in gitlab-ci K8s" was
accepted upstream as c534526e. We need to backport it here to have
it already until the next bitbake LTS version is released:

This patch adds the following variables to the allow-list, which are
used in the "IAM roles for AWS when using the GitLab chart":

- AWS_ROLE_ARN
- AWS_WEB_IDENTITY_TOKEN_FILE

These variables are set in the CI job environment and are needed to
access the sstate cache artifacts in a connected S3 bucket.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>