add support for non-default modules dir in linux-module
When building a custom kernel module, the `KBuild` file might be located in
a subdirectory. To support this use-case, set `MODULE_DIR=$(PWD)/subdir` in
the module build recipe.

Previously some sed replacements of the generated rules file where
needed to support this, in case the modules dir referenced files in
parent directories.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>