expand-on-first-boot: Preserve MBR on expansion of GPT disks
In order to keep potential hybrid partition setups intact, preserve the
legacy MBR when a GPT disk is being expanded. This is needed because we
rewrite the partitioning via sfdisk which overwrites the MBR. In the
hypbrid setup, the legacy MBR plays no role regarding the expanded
partition, thus can be simply left untouched.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>