Add support for the BeaglePlay board
This creates a minimal image for the BeaglePlay board [1], using UEFI as
boot method.

That is requiring some extra work so that the image can also be used on
SD cards: The ROM loader only supports legacy DOS partition tables, but
UEFI needs GPT. We address this by creating a hybrid image with both
partition tables in place using the new hybrid-boot bootloader source
plugin.

[1] https://www.beagleboard.org/boards/beagleplay

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>