fix dependencies of devshell tasks
In 21a0560 the local_isarapt task was introduced, which needs to run
before the devshell. By that, the dependencies of the devshell tasks
got updated as well. However, the do_prepare_build also needs to run
before the devshell, as otherwise the preparation logic is not
executed.

This commit re-introduces (adds) this dependency again.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Reviewed-by: Jan Kiszka <jan.kiszka@siemens.com>