meta-isar: starfive-visionfive2: Switch to upstream U-Boot
There are no more reasons to use the vendor tree, U-Boot 2024.01
sufficiently supports the board.

Moving upstream allows for significant simplifications in the build. We
are now building OpenSBI first and are then using U-Boot to finish all
artifacts needed for the image.

The wks file is updated accordingly, switching the GUID of the U-Boot
partition to the one upstream is recommending. Unneeded options for the
first two partitions are dropped at this chance.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>