auto disable cross for kmod build against distro kernels
Cross compiling kernel modules for distro kernels is not supported in debian.
To simplify downstream kernel module builds, we automatically turn off cross
compilation for a user-provided module when building it for a distro kernel.

This patch further replaces the hard-coded list of distro kernels with
the DISTRO_KERNELS variable to avoid code duplication.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>