fix copy out of apt cache after sbuild
During the setup of the sbuild chroot, all packages from the
global apt cache are symlinked into the sbuild chroot. After the build,
the newly downloaded packages should be copied out from the local cache
to the global apt cache. However, the current logic tried to copy out
all packages, including the symlinks. This does not work, as the
symlinks might point to non-existing locations.

This is now fixed by only copying out the newly downloaded debs (the
ones which are not a symlink).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>