dpkg-runbuild: start with clean apt-cache
The global apt-cache is imported during dpkg-runbuild and cleaned once
the function completes. This works properly on successfull builds.
However, in case of a build failure, the apt-cache dir is not cleaned
and re-read on the next build, leading to unstable builds in case
erronous data was added to the cache.

This patch adds the local apt-cache dir to the cleandirs and by that
cleans it before the function is executed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>